% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{absDiffMean}
\alias{absDiffMean}
\title{Absolute values of differences of means of succeeding elements}
\usage{
absDiffMean(y)
}
\arguments{
\item{y}{list of numerical vectors}
}
\value{
numeric(length(y) - 1)
}
\description{
Calculate mean for every element of the list, calculate the differences between succeeding means,
and return absolute values of these differences.
}
