% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computerAMEInterval.R
\name{computeAMEInterval}
\alias{computeAMEInterval}
\title{Compute AME given break points on the sample space of a feature}
\usage{
computeAMEInterval(model, data, feature, n.parts = 5, method = "ALE",
  breaks = NULL, ...)
}
\arguments{
\item{model}{[\code{\link[mlr]{WrappedModel}} | \code{any}]\cr
A model object. Can be also an object of class \code{\link[mlr]{WrappedModel}}.}

\item{data}{[\code{logical(1)}]\cr
The data set that was used to fit the model.}

\item{feature}{[\code{character(1)}]\cr
Feature name, subset of \code{colnames(data)}.}

\item{breaks}{[\code{numeric}]}
}
\description{
Compute AME given break points on the sample space of a feature
}
