% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Generate partitions with "similar" derivative}
\usage{
partition(x, y, n.parts, part.method = "CART")
}
\arguments{
\item{part.method}{c("CART", "cluster")}
}
\value{
split.points numeric(length(n.parts)-1)
}
\description{
Generate partitions with "similar" derivative
}
\section{TODOs}{

\itemize{
  \item Write tests.
  \item Alternative evaluation functions for cluster method.
}
}

