% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.AME.R
\name{plot.AME}
\alias{plot.AME}
\title{Visualize AME feature-wise}
\usage{
\method{plot}{AME}(AME, data, target, model = NA)
}
\arguments{
\item{data}{data.frame}

\item{target}{character(1) Name of target variable.}

\item{model}{}
}
\description{
Visualize AME feature-wise
}
\section{TODOS}{

\itemize{
\item Write tests (you can use examples from AME vignette)
\item Wrapped.Model / task support
\item classification support
\item Facet of different models (i.e. draw multiple AME lines for one feature)
}
}

