% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePD.R
\name{plot.PD}
\alias{plot.PD}
\title{Plot partial dependence}
\usage{
\method{plot}{PD}(x, ...)
}
\arguments{
\item{PD}{object created by \code{\link{computePD}}}
}
\value{
\code{ggplot}
}
\description{
Plot partial dependence
}
