% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeALE.R
\name{plotALE}
\alias{plotALE}
\title{Create ALE Plot}
\usage{
plotALE(ALE, derivative = FALSE)
}
\arguments{
\item{ALE}{object created by \code{\link{computeALE}}}

\item{derivative=FALSE}{If TRUE, plot ALEs, otherwise plot predictions at
interval limits.}
}
\value{
\code{ggplot}
}
\description{
Create ALE Plot
}
