% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAME.R
\name{plotAME}
\alias{plotAME}
\title{Plot AME}
\usage{
plotAME(object, data, ame, target)
}
\arguments{
\item{object}{Model object}

\item{data}{data.frame}

\item{ame}{AME object}

\item{target}{character(1) Name of target variable.}
}
\description{
Plot AME
}
\section{TODOS}{

\itemize{
\item Write tests (you can use examples from AME vignette)
\item Wrapped.Model / task support
\item classification support
\item Facet of different models (i.e. draw multiple AME lines for one feature)
}
}

