% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePD.R
\name{plotPD}
\alias{plotPD}
\title{Plot partial dependence}
\usage{
plotPD(PD)
}
\arguments{
\item{PD}{object created by \code{\link{computePD}}}
}
\value{
\code{ggplot}
}
\description{
Plot partial dependence
}
