% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrediction.R
\name{plotPrediction}
\alias{plotPrediction}
\title{Plot feature values against predictions of a model}
\usage{
plotPrediction(model, ...)
}
\arguments{
\item{feature}{Name of the feature}
}
\value{
\code{ggplot}
}
\description{
Creates univariate plots of a feature's values against model predictions and a smooth curve
using loess regression.
}
\section{TODO}{

\itemize{
  \item support multiple features at once
  \item support multiclass (see lpd_motivation for an example)
}
}

