% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeALE.R
\name{plot.ALE}
\alias{plot.ALE}
\title{Create ALE Plot}
\usage{
\method{plot}{ALE}(x, derivative = FALSE, ...)
}
\arguments{
\item{derivative}{If TRUE, plot ALEs, otherwise plot predictions at
interval limits.}

\item{ALE}{object created by \code{\link{computeALE}}}
}
\value{
\code{ggplot}
}
\description{
Create ALE Plot
}
